requirements.txt
agate==1.9.1
altair==5.5.0
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
boto3==1.40.45
botocore==1.40.45
cachetools==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.0
daff==1.4.2
dbt-adapters==1.17.2
dbt-common==1.32.0
dbt-core==1.10.13
dbt-extractor==0.6.0
dbt-protos==1.0.375
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.10.2
deepdiff==8.6.1
dlt==1.17.1
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
giturlparse==0.12.0
hexbytes==1.3.1
humanize==4.13.0
idna==3.10
importlib-metadata==8.7.0
isodate==0.6.1
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jinja2==3.1.6
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
leather==0.4.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mashumaro==3.14
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.1
narwhals==2.6.0
networkx==3.5
numpy==2.3.3
orderly-set==5.5.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
pathvalidate==3.3.1
pendulum==3.1.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
pluggy==1.6.0
ply==3.11
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.10
pydantic-core==2.33.2
pydeck==0.9.1
pygments==2.19.2
pyjwt==2.10.1
pyopenssl==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pyyaml==6.0.3
referencing==0.36.2
requests==2.32.5
requirements-parser==0.13.0
rich==14.1.0
rich-argparse==1.7.1
rpds-py==0.27.1
s3transfer==0.14.0
semver==3.0.4
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.17.4
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
sqlglot==27.20.0
sqlparse==0.5.3
streamlit==1.50.0
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
win-precise-time==1.4.2
zipp==3.23.0
